requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.12.0
openai>=1.6.1
tiktoken>=0.5.2

# Database and caching
psycopg2-binary>=2.9.9
redis>=5.0.1
sqlalchemy>=2.0.25
pgvector>=0.2.4

# Web framework
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.2
python-multipart>=0.0.6

# Document processing
pypdf2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.2
markdown>=3.5.1

# Monitoring and observability
prometheus-client>=0.19.0
grafana-api>=1.0.3

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
loguru>=0.7.2
tqdm>=4.66.1
numpy>=1.24.3
pandas>=2.1.4
scikit-learn>=1.3.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.21.1
httpx>=0.26.0
